Add <assignment_ring_view.h> with different (better?) semantics.
My thought right now today is that the version in <ring_view.h>
is unsaveable; it claims to be "non-owning", but it messes with
the lifetimes of the underlying objects, which is exactly what
it means to be "owning".

This new version is MUCH simpler. It doesn't work with objects
that don't support assignment of some sort, but its interface is
much much saner.