Add an assignment-only ring_view<T>.
From my comments on Issue #1: Basically, if ring_view is going to be
non-owning, then it shouldn't be messing with the lifetimes of the
pointed-to objects. You can't destroy something you don't own!
So that's why I'm now coming around to the opinion that pop should be
non-destructive and push should always be implemented as assignment —
we lose the ability to store non-assignable objects, but we gain sanity.

The push_back() that requires copy-assignment is enabled-if T supports
copy-assignment; and the push_back() that requires move-assignment is
enabled-if T supports move-assignment.

By adding a template type parameter `Popper`, we allow the view to
specify a (possibly mutating) "cleanup" action on popped elements.
For example, move_popper<T> moves-from the popped element, which nulls
out std::shared_ptr and std::unique_ptr objects; this preserves our
ability to use ring_view as an "owning" std::queue (where popped objects
are actually deallocated, rather than sitting around after they've been
popped). However, the default value for the `Popper` parameter is
`default_popper<T>`, which is a no-op.