Add <fixed_ring.h> along the lines of P0059R0.
Showing that fixed_ring can be implemented in terms of ring_view
plus std::array.

Unfortunately the move-constructor for fixed_ring (in this implementation)
is not trivial, because the ring_view contains pointers into the middle
of the std::array and so we have to fix those up. It would be more
efficient for the fixed_ring not to include any pointers into the array
at all: just an integer index and size.