Refactor ring_span.h and remove all noisy comments.
This is basically the "human-readable" reference implementation
as of P0059R1 plus errata. (There is no P0059R2 yet, but once there
is, I expect it to match this reference implementation.)

Significant changes:

- Remove the free functions try_push_back etc. These don't strike me
as useful enough to standardize, and they're underspecified in P0059R1.
I'm not averse to putting them back in if someone convinces me we
need them.

- Remove copy_popper. I don't think it's useful enough to standardize;
and if the user wants it, it's easy to implement.

- Rename default_popper to move_popper, now that there's no chance
of confusion between "move_popper" and "copy_popper". This popper
remains the default choice for the Popper template type argument,
in order to make ring-buffers of RAII types do the right thing
out of the box.

- Refactor all index manipulations into private member functions.
This is the first step in possibly getting rid of "%" operations
for power-of-two capacities, and/or supporting the Magic Ring Buffer
physical-memory hack.